Added menu's. 	File - close. 	Display - Foreground Colour. 	Display - Background Colour.
For both the display options, a colour picker is launched and the new colour can be selected.
The colours of the main form and the status bar are then updated.
The new colours are then saved in the config file.
These config values are used to set the Klock colours on start up.